Refactor load and pill handling: streamline server load status updates and integrate local storage updates
