+++ pages/modules/networking.js
@@ -78,5 +78,6 @@ function sendRawToAnki(data){
         xhr.send(JSON.stringify(data));
     });
 }
+window.electron_settings.sendLS(localStorage);
 
 export {tokenise, getCards, getTranslation, sendRawToAnki};
